ESQL: Refactor named writeable entries for expressions and plans (#117029)
Move writable declarations outside the core classes to avoid errors
(such as subClass.getNamedWritable()) and centralize them in a top
package class for better management.

Make all touched serialization code the same as in the initial PR,
except for the fact that LookupJoin/LookupJoinExec are missing.

(cherry picked from commit 0b74492a30f3e9c5bd366fa2c82fa637c87f5974)
