+++ x-pack/plugin/esql-core/src/main/java/org/elasticsearch/xpack/esql/core/expression/predicate/fulltext/FullTextPredicate.java
@@ -6,7 +6,6 @@
  */
 package org.elasticsearch.xpack.esql.core.expression.predicate.fulltext;
 
-import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.core.Nullable;
@@ -22,10 +21,6 @@ import java.util.Map;
 import java.util.Objects;
 
 public abstract class FullTextPredicate extends Expression {
-
-    public static List<NamedWriteableRegistry.Entry> getNamedWriteables() {
-        return List.of(MatchQueryPredicate.ENTRY, MultiMatchQueryPredicate.ENTRY, StringQueryPredicate.ENTRY);
-    }
 
     public enum Operator {
         AND,
