+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/expression/function/aggregate/AggregateFunction.java
@@ -7,7 +7,6 @@
 package org.elasticsearch.xpack.esql.expression.function.aggregate;
 
 import org.elasticsearch.TransportVersions;
-import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.xpack.esql.core.expression.Expression;
@@ -30,27 +29,6 @@ import static org.elasticsearch.xpack.esql.core.expression.TypeResolutions.Param
  * A type of {@code Function} that takes multiple values and extracts a single value out of them. For example, {@code AVG()}.
  */
 public abstract class AggregateFunction extends Function {
-    public static List<NamedWriteableRegistry.Entry> getNamedWriteables() {
-        return List.of(
-            Avg.ENTRY,
-            Count.ENTRY,
-            CountDistinct.ENTRY,
-            Max.ENTRY,
-            Median.ENTRY,
-            MedianAbsoluteDeviation.ENTRY,
-            Min.ENTRY,
-            Percentile.ENTRY,
-            Rate.ENTRY,
-            SpatialCentroid.ENTRY,
-            Sum.ENTRY,
-            Top.ENTRY,
-            Values.ENTRY,
-            // internal functions
-            ToPartial.ENTRY,
-            FromPartial.ENTRY,
-            WeightedAvg.ENTRY
-        );
-    }
 
     private final Expression field;
     private final List<? extends Expression> parameters;
