+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/expression/function/scalar/package-info.java
@@ -112,7 +112,7 @@
  *         {@link org.elasticsearch.common.io.stream.NamedWriteable#writeTo},
  *         and a deserializing constructor. Then add an {@link org.elasticsearch.common.io.stream.NamedWriteableRegistry.Entry}
  *         constant and register it. To register it, look for a method like
- *         {@link org.elasticsearch.xpack.esql.expression.function.scalar.EsqlScalarFunction#getNamedWriteables()}
+ *         {@link org.elasticsearch.xpack.esql.expression.function.scalar.ScalarFunctionWritables#getNamedWriteables()}
  *         in your function's class hierarchy. Keep going up until you hit a function with that name.
  *         Then add your new "ENTRY" constant to the list it returns.
  *     </li>
