+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/plan/physical/PhysicalPlan.java
@@ -7,7 +7,6 @@
 
 package org.elasticsearch.xpack.esql.plan.physical;
 
-import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.xpack.esql.core.tree.Source;
 import org.elasticsearch.xpack.esql.plan.QueryPlan;
 
@@ -20,34 +19,6 @@ import java.util.List;
  * PhysicalPlan = take Delta, DEN to SJC, then SJC to SFO
  */
 public abstract class PhysicalPlan extends QueryPlan<PhysicalPlan> {
-    public static List<NamedWriteableRegistry.Entry> getNamedWriteables() {
-        return List.of(
-            AggregateExec.ENTRY,
-            DissectExec.ENTRY,
-            EnrichExec.ENTRY,
-            EsQueryExec.ENTRY,
-            EsSourceExec.ENTRY,
-            EvalExec.ENTRY,
-            ExchangeExec.ENTRY,
-            ExchangeSinkExec.ENTRY,
-            ExchangeSourceExec.ENTRY,
-            FieldExtractExec.ENTRY,
-            FilterExec.ENTRY,
-            FragmentExec.ENTRY,
-            GrokExec.ENTRY,
-            HashJoinExec.ENTRY,
-            LimitExec.ENTRY,
-            LocalSourceExec.ENTRY,
-            MvExpandExec.ENTRY,
-            OrderExec.ENTRY,
-            ProjectExec.ENTRY,
-            RowExec.ENTRY,
-            ShowExec.ENTRY,
-            SubqueryExec.ENTRY,
-            TopNExec.ENTRY
-        );
-    }
-
     public PhysicalPlan(Source source, List<PhysicalPlan> children) {
         super(source, children);
     }
