+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/SerializationTestUtils.java
@@ -13,7 +13,7 @@ import org.elasticsearch.common.io.stream.BytesStreamOutput;
 import org.elasticsearch.common.io.stream.NamedWriteableAwareStreamInput;
 import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.compute.data.Block;
+import org.elasticsearch.compute.data.BlockWritables;
 import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.ExistsQueryBuilder;
 import org.elasticsearch.index.query.MatchAllQueryBuilder;
@@ -24,15 +24,11 @@ import org.elasticsearch.index.query.TermQueryBuilder;
 import org.elasticsearch.index.query.TermsQueryBuilder;
 import org.elasticsearch.index.query.WildcardQueryBuilder;
 import org.elasticsearch.test.EqualsHashCodeTestUtils;
-import org.elasticsearch.xpack.esql.core.expression.Attribute;
 import org.elasticsearch.xpack.esql.core.expression.Expression;
-import org.elasticsearch.xpack.esql.core.expression.NamedExpression;
-import org.elasticsearch.xpack.esql.expression.function.UnsupportedAttribute;
-import org.elasticsearch.xpack.esql.expression.function.aggregate.AggregateFunction;
-import org.elasticsearch.xpack.esql.expression.function.fulltext.FullTextFunction;
-import org.elasticsearch.xpack.esql.expression.function.scalar.EsqlScalarFunction;
+import org.elasticsearch.xpack.esql.expression.ExpressionWritables;
 import org.elasticsearch.xpack.esql.io.stream.PlanStreamInput;
 import org.elasticsearch.xpack.esql.io.stream.PlanStreamOutput;
+import org.elasticsearch.xpack.esql.plan.PlanWritables;
 import org.elasticsearch.xpack.esql.plan.logical.LogicalPlan;
 import org.elasticsearch.xpack.esql.plan.physical.PhysicalPlan;
 import org.elasticsearch.xpack.esql.querydsl.query.SingleValueQuery;
@@ -115,17 +111,9 @@ public class SerializationTestUtils {
         entries.add(new NamedWriteableRegistry.Entry(QueryBuilder.class, RegexpQueryBuilder.NAME, RegexpQueryBuilder::new));
         entries.add(new NamedWriteableRegistry.Entry(QueryBuilder.class, ExistsQueryBuilder.NAME, ExistsQueryBuilder::new));
         entries.add(SingleValueQuery.ENTRY);
-        entries.addAll(Attribute.getNamedWriteables());
-        entries.add(UnsupportedAttribute.ENTRY);
-        entries.addAll(NamedExpression.getNamedWriteables());
-        entries.add(UnsupportedAttribute.NAMED_EXPRESSION_ENTRY);
-        entries.addAll(Expression.getNamedWriteables());
-        entries.addAll(EsqlScalarFunction.getNamedWriteables());
-        entries.addAll(AggregateFunction.getNamedWriteables());
-        entries.addAll(Block.getNamedWriteables());
-        entries.addAll(LogicalPlan.getNamedWriteables());
-        entries.addAll(FullTextFunction.getNamedWriteables());
-        entries.addAll(PhysicalPlan.getNamedWriteables());
+        entries.addAll(ExpressionWritables.getNamedWriteables());
+        entries.addAll(PlanWritables.getNamedWriteables());
+        entries.addAll(BlockWritables.getNamedWriteables());
         return new NamedWriteableRegistry(entries);
     }
 }
