+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/plan/logical/AbstractLogicalPlanSerializationTests.java
@@ -8,13 +8,11 @@
 package org.elasticsearch.xpack.esql.plan.logical;
 
 import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
-import org.elasticsearch.compute.data.Block;
-import org.elasticsearch.xpack.esql.core.expression.Attribute;
-import org.elasticsearch.xpack.esql.core.expression.Expression;
-import org.elasticsearch.xpack.esql.core.expression.NamedExpression;
+import org.elasticsearch.compute.data.BlockWritables;
 import org.elasticsearch.xpack.esql.core.tree.Node;
-import org.elasticsearch.xpack.esql.expression.function.aggregate.AggregateFunction;
+import org.elasticsearch.xpack.esql.expression.ExpressionWritables;
 import org.elasticsearch.xpack.esql.plan.AbstractNodeSerializationTests;
+import org.elasticsearch.xpack.esql.plan.PlanWritables;
 import org.elasticsearch.xpack.esql.plan.logical.local.LocalRelationSerializationTests;
 
 import java.util.ArrayList;
@@ -32,12 +30,10 @@ public abstract class AbstractLogicalPlanSerializationTests<T extends LogicalPla
     @Override
     protected final NamedWriteableRegistry getNamedWriteableRegistry() {
         List<NamedWriteableRegistry.Entry> entries = new ArrayList<>();
-        entries.addAll(LogicalPlan.getNamedWriteables());
-        entries.addAll(AggregateFunction.getNamedWriteables());
-        entries.addAll(Expression.getNamedWriteables());
-        entries.addAll(Attribute.getNamedWriteables());
-        entries.addAll(Block.getNamedWriteables());
-        entries.addAll(NamedExpression.getNamedWriteables());
+        entries.addAll(PlanWritables.logical());
+        entries.addAll(ExpressionWritables.aggregates());
+        entries.addAll(ExpressionWritables.allExpressions());
+        entries.addAll(BlockWritables.getNamedWriteables());
         return new NamedWriteableRegistry(entries);
     }
